Did away with package managers for the time being
This project is, more than anything, a tool for me to learn.
One of the biggest takeaways so far is that the state of C++
package management is absolutely abysmal. In a perfect world,
I would be able to use a package manager like VCPKG, Conan, etc.
to handle all external headers, dependencies, and libraries.
Unfortunately, however, I do not live in that perfect world.

Instead, headers are now located in the lib directory, and
CMake handles finding dynamic libraries, which must be present
on the build system.

With the next release, I will be providing pre-built binaries
in the releases tab, on Homebrew, and hopefully through PPA
so that the end user does not have to worry to much about this.

Okay, rant over...