Updated CI/CD pipeline to use Docker
The pipeline had gone out of date and was failing. The exact failure had to do with APT failing to fetch `libinput10_1.25.0-1ubuntu3.1_amd64.deb`. Regardless, the pipeline had as a whole fallen out of date anyway. I moved most of the configuration and build process into a Dockerfile, which GitHub Actions will build and run to perform unit testing.