Updated CPM
Prior to this commit, the `cmake` directory held a file containing source code for the CMake Package Manager (CPM). The version of CPM contained in the file is now incompatible with newer versions of CMake. I have updated this project to use the latest CPM version. I also replaced the file with a statement in the project's `CMakeLists.txt` that fetches CPM from the Internet during the build process. This will make it easier to update CPM in the future.