Updated unit test workflow to use Dockerfile (#62)
* Updated CI/CD pipeline to use Docker

The pipeline had gone out of date and was failing. The exact failure had to do with APT failing to fetch `libinput10_1.25.0-1ubuntu3.1_amd64.deb`. Regardless, the pipeline had as a whole fallen out of date anyway. I moved most of the configuration and build process into a Dockerfile, which GitHub Actions will build and run to perform unit testing.

* Updated CPM

Prior to this commit, the `cmake` directory held a file containing source code for the CMake Package Manager (CPM). The version of CPM contained in the file is now incompatible with newer versions of CMake. I have updated this project to use the latest CPM version. I also replaced the file with a statement in the project's `CMakeLists.txt` that fetches CPM from the Internet during the build process. This will make it easier to update CPM in the future.